+++ astropy/convolution/tests/test_kernel_class.py
@@ -204,7 +204,7 @@ class TestKernels(object):
 
     def test_custom_1D_kernel(self):
         """
-        Check if CustomKernel against Box1DKernel.
+        Check CustomKernel against Box1DKernel.
         """
         #Define one dimensional array:
         array = np.ones(5)
@@ -218,7 +218,7 @@ class TestKernels(object):
 
     def test_custom_2D_kernel(self):
         """
-        Check if CustomKernel against Box1DKernel.
+        Check CustomKernel against Box2DKernel.
         """
         #Define one dimensional array:
         array = np.ones((5, 5))
@@ -246,9 +246,29 @@ class TestKernels(object):
                                [1, 1, 1]])
         assert custom.is_bool == True
 
+    def test_custom_1D_kernel_zerosum(self):
+        """
+        Check if CustomKernel works when the input array/list
+        sums to zero.
+        """
+        custom = CustomKernel([-2, -1, 0, 1, 2])
+        assert custom.truncation == 0.
+        assert custom.normalization == np.inf
+
+    def test_custom_2D_kernel_zerosum(self):
+        """
+        Check if CustomKernel works when the input array/list
+        sums to zero.
+        """
+        custom = CustomKernel([[0, -1, 0],
+                               [-1, 4, -1],
+                               [0, -1, 0]])
+        assert custom.truncation == 0.
+        assert custom.normalization == np.inf
+
     def test_custom_kernel_odd_error(self):
         """
-        Check if CustomKErnel raises if the array size is odd.
+        Check if CustomKernel raises if the array size is odd.
         """
         with pytest.raises(KernelSizeError):
             custom = CustomKernel([1, 1, 1, 1])
@@ -370,4 +390,4 @@ class TestKernels(object):
         assert np.all([_ % 2 != 0 for _ in kernel_2D.shape])
         assert kernel_2D.array.sum() == 1.
 
-    
+