+++ astropy/time/tests/test_comparisons.py
@@ -1,4 +1,7 @@
+
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+import operator
+
 import numpy as np
 
 from ...tests.helper import pytest
@@ -11,6 +14,19 @@ class TestTimeComparisons():
     def setup(self):
         self.t1 = Time(np.arange(49995, 50005), format='mjd', scale='utc')
         self.t2 = Time(np.arange(49000, 51000, 200), format='mjd', scale='utc')
+
+    def test_miscompares(self):
+        t1 = Time('J2000', scale='utc')
+        for op, op_str in ((operator.eq, '=='),
+                           (operator.ge, '>='),
+                           (operator.gt, '>'),
+                           (operator.ne, '!='),
+                           (operator.le, '<='),
+                           (operator.lt, '<')):
+            with pytest.raises(OperandTypeError) as err:
+                op(t1, None)
+            assert str(err).endswith("Unsupported operand type(s) for {0}: 'Time' and 'NoneType'"
+                                     .format(op_str))
 
     def test_time(self):
         t1_lt_t2 = self.t1 < self.t2