+++ astropy/wcs/wcs.py
@@ -1176,11 +1176,6 @@ naxis kwarg.
 
         {1}
 
-        tolerance : float, optional
-            Tolerance of solution. Iteration terminates when the iterative
-            solver estimates that the true solution is within this many pixels
-            current estimate. Default value is 1e-6 (pixels).
-
         Returns
         -------
 
@@ -1273,6 +1268,11 @@ naxis kwarg.
             two-argument form must be used.
 
         {1}
+
+        tolerance : float, optional
+            Tolerance of solution. Iteration terminates when the iterative
+            solver estimates that the true solution is within this many pixels
+            current estimate. Default value is 1e-6 (pixels).
 
         Returns
         -------
@@ -2063,10 +2063,13 @@ def validate(source):
                 hdu.header, relax=True, fix=False, _do_set=False)
 
         for wcs in wcses:
-            wcs_results = _WcsValidateWcsResult(wcs.wcs.name)
+            wcs_results = _WcsValidateWcsResult(wcs.wcs.alt)
             hdu_results.append(wcs_results)
 
-            del __warningregistry__
+            try:
+                del __warningregistry__
+            except NameError:
+                pass
 
             with warnings.catch_warnings(record=True) as warning_lines:
                 warnings.resetwarnings()
@@ -2075,7 +2078,7 @@ def validate(source):
 
                 try:
                     WCS(hdu.header,
-                        key=wcs.wcs.name or ' ',
+                        key=wcs.wcs.alt or ' ',
                         relax=True, fix=True)
                 except WcsError as e:
                     wcs_results.append(str(e))